% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM data}
\format{
a data frame with 16 rows and 2 variables
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{fire damage done in thousands of dollars}
  ...
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A dataset of fire damage and distance
}
\keyword{datasets}
