% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample size}

\item{fun}{function used}

\item{alpha}{confidence level}

\item{cx}{constant value}

\item{...}{allows for additional arguments}
}
\value{
a histogram of bootstrap sample statistics
}
\description{
myboot2
}
\examples{
\dontrun{myboot2(iter=10000, x=sam, fun="sd", alpha=0.2, xlab="standard deviation",col="#7D5BA6")}
}
