% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{size}

\item{iter}{number of iterations}

\item{a}{lower limit}

\item{b}{upper limit}
}
\value{
a histogram
}
\description{
myclt
}
\examples{
\dontrun{w=myclt(n=50,iter=10000,a=5,b=10)}

}
