% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{mymlnorm}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{sample vector}

\item{mu}{mean}

\item{sig}{standard deviation}

\item{...}{allows for additional arguments}
}
\value{
a contour plot of the likelihood function
}
\description{
mymlnorm
}
\examples{
\dontrun{mymlnorm(
x=c(5,7,7,8,10),
mu=seq(5,10,length=1000),
sig=seq(0.1,4,length=1000),
lwd=2,labcex=1)}

}
