% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
a normal distribution curve
}
\description{
myncurve
}
\examples{
\dontrun{myncurve(mu=4, sigma=3)}

}
