% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{mypvalue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{sample value}

\item{xmax}{max value shown in plot}

\item{n}{size of sample}

\item{alpha}{significance level}
}
\value{
a plot with talpha values. It also returns the p-value
}
\description{
mypvalue
}
\examples{
\dontrun{mypvalue(tcalc,n=30,alpha=0.05)}

}
