% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{myread}
\usage{
myread(csv)
}
\arguments{
\item{csv}{excel file}
}
\value{
csv file that can be read in
}
\description{
myread
}
\examples{
\dontrun{fin.df=myread("FINTUBES.csv")}

}
