% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on a flight}

\item{gamma}{the probability of overbooking}

\item{p}{the probability a passenger will show}
}
\value{
a list of nc, nd, N, p, gamma, and a plot
}
\description{
ntickets
}
\examples{
\dontrun{ntickets(N=200, gamma=0.02, p=0.95)}

}
